[HUMAN App] refactor: hcaptcha-labeling structure (#3041)
