+++ packages/apps/human-app/frontend/src/modules/worker/hcaptcha-labeling/components/index.ts
@@ -1,0 +1,3 @@
+export * from './user-stats-accordion';
+export * from './user-stats-details';
+export * from './user-stats-drawer';
