+++ packages/apps/human-app/frontend/src/modules/worker/hcaptcha-labeling/components/user-stats-details.tsx
@@ -1,17 +1,17 @@
 import { Divider, Grid, Typography } from '@mui/material';
 import { t } from 'i18next';
 import { RefreshIcon } from '@/shared/components/ui/icons';
-import type { HCaptchaUserStatsSuccess } from '@/modules/worker/services/hcaptcha-user-stats';
-import { useColorMode } from '@/shared/contexts/color-mode';
 import { onlyDarkModeColor } from '@/shared/styles/dark-color-palette';
+import { useColorMode } from '@/shared/contexts/color-mode';
+import { type HCaptchaUserStatsSuccess } from '../hooks/use-hcaptcha-user-stats';
 
 export function UserStatsDetails({
   stats,
   refetch,
-}: {
+}: Readonly<{
   stats: HCaptchaUserStatsSuccess;
   refetch: () => void;
-}) {
+}>) {
   const { colorPalette, isDarkMode } = useColorMode();
   const statsColor = isDarkMode
     ? onlyDarkModeColor.additionalTextColor
