+++ packages/apps/human-app/frontend/src/modules/worker/hcaptcha-labeling/hcaptcha-labeling.page.tsx
@@ -8,13 +8,7 @@ import { useIsMobile } from '@/shared/hooks/use-is-mobile';
 import { env } from '@/shared/env';
 import { breakpoints } from '@/shared/styles/breakpoints';
 import { Counter } from '@/shared/components/ui/counter';
-import { useHCaptchaUserStats } from '@/modules/worker/services/hcaptcha-user-stats';
-import {
-  PageCardError,
-  PageCardLoader,
-} from '@/shared/components/ui/page-card';
 import { getErrorMessageForError } from '@/shared/errors';
-import { useDailyHmtSpent } from '@/modules/worker/services/daily-hmt-spent';
 import { getTomorrowDate } from '@/shared/helpers/date';
 import { useSolveHCaptchaMutation } from '@/modules/worker/services/solve-hcaptcha';
 import { useAuthenticatedUser } from '@/modules/auth/hooks/use-authenticated-user';
@@ -22,6 +16,11 @@ import { useHCaptchaLabelingNotifications } from '@/modules/worker/hooks/use-hca
 import { useColorMode } from '@/shared/contexts/color-mode';
 import { onlyDarkModeColor } from '@/shared/styles/dark-color-palette';
 import { routerPaths } from '@/router/router-paths';
+import {
+  PageCardLoader,
+  PageCardError,
+} from '@/shared/components/ui/page-card';
+import { useHCaptchaUserStats, useDailyHmtSpent } from './hooks';
 
 export function HcaptchaLabelingPage() {
   const { colorPalette, isDarkMode } = useColorMode();
