+++ packages/apps/human-app/frontend/src/modules/worker/hcaptcha-labeling/hooks/use-hcaptcha-user-stats.ts
@@ -25,7 +25,7 @@ const hcaptchaUserStatsSchema = z.object({
 
 export type HCaptchaUserStatsSuccess = z.infer<typeof hcaptchaUserStatsSchema>;
 
-export async function getHCaptchaUsersStats() {
+async function getHCaptchaUsersStats() {
   return apiClient(apiPaths.worker.hCaptchaUserStats.path, {
     authenticated: true,
     successSchema: hcaptchaUserStatsSchema,
