+++ packages/apps/human-app/frontend/src/modules/worker/hcaptcha-labeling/index.ts
@@ -1,0 +1,4 @@
+export * from './enable-labeler.page';
+export * from './hcaptcha-labeling.page';
+export * from './components/user-stats-accordion';
+export * from './components/user-stats-drawer';
